Pass size_t to span functions: chrome/
In preparation for upgrading various span args from size_t to
StrictNumeric<size_t>, this fixes type mismatches that would cause
compile errors with that change.

Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6020621
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1382641}

+++ chrome/browser/web_applications/preinstalled_web_apps/preinstalled_web_app_definition_utils.cc
@@ -5,6 +5,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "chrome/browser/web_applications/preinstalled_web_apps/preinstalled_web_app_definition_utils.h"
 
+#include "base/numerics/safe_conversions.h"
 #include "chrome/browser/browser_process.h"
 #include "third_party/icu/source/common/unicode/localematcher.h"
 #include "ui/base/resource/resource_bundle.h"
@@ -56,7 +57,7 @@ const char* GetTranslatedName(const char* utf8_default_name,
   if (best_index == -1)
     return utf8_default_name;
 
-  return translations[best_index].utf8_translation;
+  return translations[base::checked_cast<size_t>(best_index)].utf8_translation;
 }
 
 }  // namespace web_app
